+++ ChangeLog.md
@@ -15,6 +15,13 @@ and this project adheres to
 
 - ...
 
+### Fixed
+
+- fixed a potential non closing of a FileOutputStream with function json/spit
+- dropping a file to the REPL that has special characters (space, asteriks, 
+  ...) in the filename. The underlying OS shell is escaping these characters.
+  E.g.:  "test\ 1.venice", "test\?1.venice"
+
 
 
 ## [1.9.14] - 2021-04-04